Moved classes to better package
