+++ src/main/groovy/org/mockito/release/internal/gradle/configuration/LazyValidator.java
@@ -1,4 +1,4 @@
-package org.mockito.release.internal.gradle.util;
+package org.mockito.release.internal.gradle.configuration;
 
 import org.gradle.api.Action;
 import org.gradle.api.GradleException;
